fixed navbar issue (bold text), also general formatting and organizational things
Fixed current subpage (bold text) altering width of containers in the navbar. Set exact width values based on the width of each element with bold text.

Renamed custom classes to better reflect that they are custom in order to avoid confusion with the many Bootstrap classes that are used.

Changed navbar to be fixed at the top of the page (should not disappear while scrolling down). Added spacing for the content of the body to avoid overlapping with the navbar.

Renamed styles.css to shared-styles.css for better file organization.