Removed projects from coding page - need to overhaul, added SEO tags
Removed all projects listed in coding.html to prepare for overhaul of that page.

Added SEO tags to each page. Currently all are set to noindex, nofollow while the site is still in development. The plan is to make all of them (except 404.html) indexable, followable, etc. once the site is complete, but that may change. Also added canonical link tags.