Updated footer
Removed GitHub Pages and Bootstrap links. Replaced with GitHub repo link and privacy policy link.

Removing the two links should be more in line with how the repo (specifically README.md, which is automatically displayed by GitHub when viewing the repo online) is organized. The repo link is added to make it more presently available. I will likely add a privacy policy as the Adobe PDF Viewer appears to collect user data.

Also reduced border size of the navbar logo to 2px, which is purely a cosmetic change.