Testing stylization for resumes...
Adobe's PDF Embed API *requires* an API key which has to be used from the registered site. Soooo that means I have to upload it to the live site. Every time. So there's probably going to be a lot of commits coming up because Adobe really didn't think this through.