Added privacy policy page skeleton, reverted to fixed dimensions for resumes
Added privacy policy page skeleton. It currently just contains the navbar and footer (and everything in the head element). Once I finish developing the site, I will write the privacy policy in order to ensure that I don't miss anything.

Reverted to fixed dimensions for resume containers in order to display properly. This will need to be modified in the future to accommodate smaller, different aspect ratio screens such as on phones.

Some minor stylization changes for consistency's sake. Also reverted navbar logo border to 3px - that may be changed again in the future.