Added script to rotate navbar menu button when clicked/pressed
Added navbar-button.js to rotate the navbar menu button when it is clicked/pressed. Modified styles.css to define the rotation animations and also remove the navbar menu button's border.

Modified privacy.html and README.md to reflect this change.

Modified all HTML files to import the script when each page is loaded.

Minor stylization change in index.html - increased size of contact info header to be consistent with the sections on the rest of the page.