Made navbar dropdown links take up (almost) the full width of the page
QoL feature that I wanted - this means that in mobile format, you don't have to click/press the word itself in the navbar dropdown menu, you can just click/press anywhere horizontal of it.

Reverts to original width when the dropdown menu is closed to help maintain on-the-fly navbar responsiveness.