Adjusted navbar scroll feature to accommodate mouse movements
navbar-scroll.js - the navbar now appears when the mouse is over the navbar element. It subsequently disappears when the mouse moves away from the element. However, with respect to mouse movement, the navbar only disappears if the new mouse location is visually lower than the previous mouse location. This allows for making the navbar appear via scrolling up, even when the mouse is not at the top, and keeping it visible while moving the mouse to it.