Added dark mode button w/ animations, added cookies for remembering preferences
Added dark mode button animations. Works the same as navbar re: scroll position and mouse position, except mouse position relies on being on the left side of the screen instead of the top of the screen.

Added cookies to remember dark mode preferences. Cookies are deleted if dark mode is toggled off.