Bug fix - fixed issue re: previous few commits
Semaphore idea was correct, but the scroll detector is a little too delayed, so the way I implemented the semaphores didn't quite work. Reworked it so that, if scrolling is detected during a switch to/from dark mode, then the scroll detector block will be responsible for resetting the semaphore.