Bug fix - dark mode button visibility
Dark mode button disappeared while scrolling down, even if the mouse was in the "active" area on the left side of the screen. Fixed so that scrolling will only make the dark mode button disappear if the mouse is NOT on the left side of the screen.

Kept functionality of disappearing on scroll regardless of mouse position for mobile format.