Centered text on navbar for non-mobile format
Text for navbar options is now justified to the center. This makes it so that, when switching to a new page (and that new page being bolded in the navbar), it does not bold to the right anymore, now it bolds to the left and the right. This is very subtle and I don't know how to explain it better than that, but I promise that (for me) it is better and it works.

Added a JS file to make it so that this does not affected the dropdown navbar on mobile format.