Updated privacy.html and README.md to reflect Clicky.
I'm thinking I'm actually going to keep both Clicky and Plausible in the long-term, or at least not too short of a term. Reflected this in privacy.html and README.md .Also reformatted the Clicky and Plausible script tags in each html file to be a little bit cleaner visually while coding.