Added skeleton sitemap.html page. Modified footers to add sitemap link.
Also modified the comment before the dark mode preference deletion script tag on each page to be more accurate. I changed the script itself a while ago and forgot to update that comment. Also modified that script again so that it breaks out of the for loop upon successful cookie deletion. This would only have performance benefits if there are for some reason a huge amount of cookies (there should only ever be a max of 1 though), but I did it for consistency's sake I guess.