Added responsiveness to left vs. center aligned navbar text and random image.
Left vs. center aligned navbar item text is now responsive and does not require the page to be reloaded if the viewport size changes.

Similarly, the random image in index.html is now also responsive and does not require the page to be reloaded if the viewport size changes. If the viewport size is updated, the page will either 1) use the same image file that was initially chosen or 2) if the page started in mobile format (meaning no image file was chosen), then a random image file will be chosen and, unless the page is reloaded, it will continue by following rule 1).